+++ packages/admin-date-picker/src/DateRangePicker.tsx
@@ -1,4 +1,4 @@
-import * as moment from "moment";
+import moment from "moment";
 import * as React from "react";
 import { DateRangePicker as AirBNBDateRangePicker } from "react-dates";
 import { FieldRenderProps } from "react-final-form";