+++ packages/admin-stories/src/admin/form/RangeInput.tsx
@@ -1,4 +1,5 @@
-import { Field, FinalFormRangeInput } from "@comet/admin";
+import { Field, FinalFormRangeInput, FormPaper, Toolbar, ToolbarTitleItem } from "@comet/admin";
+import { Box, Button, Typography } from "@material-ui/core";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 import { Form } from "react-final-form";
@@ -37,24 +38,120 @@ const Thumb = styled.div`
 
 function Story() {
     return (
-        <div style={{ width: "300px" }}>
-            <Form
-                onSubmit={(values) => {
-                    // values
-                }}
-                initialValues={{ price: { min: 0, max: 100 } }}
-                render={({ handleSubmit, values }) => (
-                    <Field
-                        component={FinalFormRangeInput}
-                        name="price"
-                        min={0}
-                        max={100}
-                        endAdornment={<span>€</span>}
-                        sliderProps={{ ThumbComponent: Thumb }}
-                    />
-                )}
-            />
-        </div>
+        <>
+            <Box marginBottom={4}>
+                <Toolbar>
+                    <ToolbarTitleItem>Final Form Range Input</ToolbarTitleItem>
+                </Toolbar>
+            </Box>
+            <div style={{ width: 400 }}>
+                <Box marginBottom={4}>
+                    <FormPaper variant="outlined">
+                        <Typography variant="h3" gutterBottom>
+                            with initialValues
+                        </Typography>
+                        <Form
+                            onSubmit={(values) => {
+                                // values
+                            }}
+                            initialValues={{ price: { min: 0, max: 100 } }}
+                            render={({ handleSubmit, values, form, initialValues }) => (
+                                <>
+                                    <Field
+                                        component={FinalFormRangeInput}
+                                        name="price"
+                                        min={0}
+                                        max={100}
+                                        endAdornment={<span>€</span>}
+                                        sliderProps={{ ThumbComponent: Thumb }}
+                                    />
+                                    <Button
+                                        variant="contained"
+                                        color="primary"
+                                        onClick={() => {
+                                            form.reset();
+                                        }}
+                                    >
+                                        Reset
+                                    </Button>
+                                    <pre>{JSON.stringify(values, undefined, 2)}</pre>
+                                </>
+                            )}
+                        />
+                    </FormPaper>
+                </Box>
+
+                <Box marginBottom={4}>
+                    <FormPaper variant="outlined">
+                        <Typography variant="h3" gutterBottom>
+                            without initialValues
+                        </Typography>
+                        <Form
+                            onSubmit={(values) => {
+                                // values
+                            }}
+                            render={({ values, form }) => (
+                                <>
+                                    <Field
+                                        component={FinalFormRangeInput}
+                                        name="price"
+                                        min={0}
+                                        max={150}
+                                        endAdornment={<span>€</span>}
+                                        sliderProps={{ ThumbComponent: Thumb }}
+                                    />
+                                    <Button
+                                        variant="contained"
+                                        color="primary"
+                                        onClick={() => {
+                                            form.reset();
+                                        }}
+                                    >
+                                        Reset
+                                    </Button>
+                                    <pre>{JSON.stringify(values, undefined, 2)}</pre>
+                                </>
+                            )}
+                        />
+                    </FormPaper>
+                </Box>
+                <Box marginBottom={4}>
+                    <FormPaper variant="outlined">
+                        <Typography variant="h3" gutterBottom>
+                            with initialValues and different MinMax Values
+                        </Typography>
+                        <Form
+                            onSubmit={(values) => {
+                                // values
+                            }}
+                            initialValues={{ price: { min: 50, max: 80 } }}
+                            render={({ values, form, initialValues }) => (
+                                <>
+                                    <Field
+                                        component={FinalFormRangeInput}
+                                        name="price"
+                                        min={20}
+                                        max={150}
+                                        endAdornment={<span>€</span>}
+                                        sliderProps={{ ThumbComponent: Thumb }}
+                                    />
+                                    <Button
+                                        variant="contained"
+                                        color="primary"
+                                        onClick={() => {
+                                            form.reset();
+                                        }}
+                                    >
+                                        Reset
+                                    </Button>
+                                    <pre>{JSON.stringify(values, undefined, 2)}</pre>
+                                </>
+                            )}
+                        />
+                    </FormPaper>
+                </Box>
+            </div>
+        </>
     );
 }