+++ packages/admin-theme/src/AdminOverrides/finalFormSaveCancelButtonsLegacy.ts
@@ -1,8 +1,0 @@
-import { CometAdminCometAdminFinalFormSaveCancelButtonsLegacyClassKeys } from "@comet/admin";
-import { StyleRules } from "@material-ui/styles/withStyles";
-
-export const getFinalFormSaveCancelButtonsLegacyOverrides = (): StyleRules<{}, CometAdminCometAdminFinalFormSaveCancelButtonsLegacyClassKeys> => ({
-    root: {},
-    cancelButton: {},
-    saveButton: {},
-});