+++ packages/admin-theme/src/AdminProps/menu.ts
@@ -1,7 +1,0 @@
-import { MenuThemeProps } from "@comet/admin";
-
-export const getMenuProps = (): MenuThemeProps => ({
-    permanentDrawerPaperProps: {
-        elevation: 2,
-    },
-});