+++ packages/admin-theme/src/AdminProps/toolbar.tsx
@@ -1,5 +1,0 @@
-import { ToolbarThemeProps } from "@comet/admin";
-
-export const getToolbarProps = (): ToolbarThemeProps => ({
-    elevation: 1,
-});