+++ packages/admin-theme/src/AdminSelectOverrides/getAdminSelectOverrides.ts
@@ -1,7 +1,0 @@
-import { Overrides } from "@material-ui/core/styles/overrides";
-
-import { getSelectOverrides } from "./select";
-
-export const getAdminSelectOverrides = (): Overrides => ({
-    CometAdminSelect: getSelectOverrides(),
-});