+++ packages/admin/src/snackbar/UndoSnackbar.tsx
@@ -2,11 +2,11 @@ import { Button, Slide, Snackbar, SnackbarProps } from "@material-ui/core";
 import { TransitionProps } from "@material-ui/core/transitions";
 import * as React from "react";
 import { FormattedMessage } from "react-intl";
+import * as UUID from "uuid";
 
 import { useSnackbarApi } from "./SnackbarProvider";
 
 export interface UndoSnackbarProps<Payload extends unknown> extends Omit<SnackbarProps, "action"> {
-    key: React.Key;
     message: React.ReactNode;
     onUndoClick: (payload?: Payload) => void;
     payload?: Payload;
@@ -14,6 +14,7 @@ export interface UndoSnackbarProps<Payload extends unknown> extends Omit<Snackba
 
 export const UndoSnackbar = <Payload extends unknown>({ onUndoClick, payload, ...props }: UndoSnackbarProps<Payload>) => {
     const snackbarApi = useSnackbarApi();
+    const uuid = React.useRef(UUID.v4());
 
     const onClick = () => {
         snackbarApi.hideSnackbar();
@@ -22,6 +23,7 @@ export const UndoSnackbar = <Payload extends unknown>({ onUndoClick, payload, ..
 
     return (
         <Snackbar
+            key={uuid.current}
             anchorOrigin={{ vertical: "bottom", horizontal: "left" }}
             autoHideDuration={5000}
             action={