+++ packages/admin/src/themeAugmentation.ts
@@ -10,6 +10,8 @@ import { CometAdminMenuCollapsibleItemClassKeys, MenuCollapsibleItemThemeProps }
 import { CometAdminMenuItemClassKeys } from "./mui/menu/Item";
 import { MenuThemeProps } from "./mui/menu/Menu";
 import { CometAdminMenuClassKeys } from "./mui/menu/Menu.styles";
+import { CometAdminFilterBarClassKeys } from "./table/filterbar/FilterBar";
+import { CometAdminFilterBarPopoverFilterClassKeys, FilterBarPopoverFilterThemeProps } from "./table/filterbar/FilterBarPopoverFilter";
 
 declare module "@material-ui/core/styles/overrides" {
     interface ComponentNameToClassKey {
@@ -24,6 +26,8 @@ declare module "@material-ui/core/styles/overrides" {
         CometAdminFormPaper: CometAdminFormPaperKeys;
         CometAdminFormSection: CometAdminFormSectionKeys;
         CometAdminFinalFormRangeInput: CometAdminFinalFormRangeInputClassKeys;
+        CometAdminFilterBar: CometAdminFilterBarClassKeys;
+        CometAdminFilterBarPopoverFilter: CometAdminFilterBarPopoverFilterClassKeys;
     }
 }
 
@@ -35,5 +39,6 @@ declare module "@material-ui/core/styles/props" {
         CometAdminMenuCollapsibleItem: MenuCollapsibleItemThemeProps;
         CometAdminMasterLayout: MasterLayoutThemeProps;
         CometAdminErrorBoundary: ErrorBoundaryThemeProps;
+        CometAdminFilterBarPopoverFilter: FilterBarPopoverFilterThemeProps;
     }
 }